Add yaem as dependency
I chose a peer dependency because yaem provides an event manager which
is a singleton. Depending on yaem normally would add a risk to have several
instances of event manager: one for widgetjs and one for the package
depending on widgetjs and on yaem.

Also add yaem as dev dependencies for unit tests.