Decompose Widget.renderRootOn() to facilitate overriding
With this decomposition, widgets can just override `_makeRootTagOn()`
when they want to control the root tag and not get
"widgetjs-widget". The following describes the problem a bit more.

Sometimes, you don't want a widget to render a `<widgetjs-widget>`
node in the DOM. This happens for example when a widget is responsible
for rendering a cell in a table. A `<tr>` can't have a
`<widgetjs-widget>` as child. In this case, you want the widget to
render a `<td>` directly.

There are several ways to avoid having a `<widgetjs-widget>` node in
the rendering of a widget:

- You can override `renderOn(html)`. This works great but you have to
make sure to only write 1 root node. The API makes that a very hidden
constraint. You also have to make sure that `asJQuery()` will keep
working (in practice by adding the widget's id as attribute to the
root node).

- You can also override `_renderRootOn(html)`. The method name makes
it a little more explicit that you should only render 1 root node but
you still have to pay attention that `asJQuery()` will keep
working. Also, this solution makes it very unclear that something has
to be returned.

This commit introduces a third, recommended, way: let the subclasses
override `_makeRootTagOn(html)`. The method name starting with `make`
makes it explicit that the method must return one and only one
node. This solution doesn't require thinking about `asJQuery()` as
this is handled automatically.