15549487: Remove _initialize methods from Widget2
It is a good practice to prefer composition over inheritance. And
since we don't make widgets inheriting from other widgets anymore, we
don't have to separate the initialization of instance variables and
the initialization of subwidgets.

This is removing _initialize and _initializeSubwidgets in favor of
always using the object constructor instead.