40698660: Make Widget.asJQuery() find nodes in specially-crafted shadow trees
In the previous version, widget.asJQuery() would only return something
useful if the widget is directly attached to the document. If the
widget is within a shadow tree, asJQuery() wouldn't work.

This commit makes asJQuery() work even if widget is within a shadow
tree if and only if:

- this tree is attached to a host node with widgetjs-shadow=document,

- and there is only one such node in the DOM

This limitation on the host avoids impacting too much the performance.